feat: add token authentication

+++ src/graphql/note/services/mutation/delete-note-service.ts
@@ -1,4 +1,5 @@
 import { INotesRepository } from "../../../../repositories/notes-repository";
+import { UserContext } from "../../../user/resolver";
 
 export interface DeleteNoteArgs {
   id: string;
@@ -6,8 +7,11 @@ export interface DeleteNoteArgs {
 
 export async function deleteNoteService(
   args: DeleteNoteArgs,
-  repository: INotesRepository
+  repository: INotesRepository,
+  { user }: UserContext
 ) {
+  if (!user) throw new Error("You are not authenticated.");
+
   const deleted = await repository.deleteNote(Number(args.id));
 
   if (!deleted) throw new Error("Error while deleting a note.");
